add getAllVehicles and addVehicle method to get and add new data

+++ tra.task/src/main/java/com/example/tra/task/Entities/Vehicle.java
@@ -10,5 +10,5 @@ import lombok.NoArgsConstructor;
 public class Vehicle {
     private String vehicleId;
     private String vehicleModel;
-    private double rentalPricePerDay;
+    private Integer rentalPricePerDay;
 }
